refactor: increase the round timeout to account for backoff
The max value that the suggested sleep time can be after introducing the backoff logic is 32 (backoff_factor ** max_retries) with the current configuration values. Therefore, we set the round timeout to 35 to also allow for other operations to be performed during a round that uses the maximum sleep time suggestion.