doc: make set up step explicit
Make set up step explicit in the demo instructions so that it indicates the specific version of Open Autonomy that the service currently supports.

This is done in order to avoid situations where the Open Autonomy framework has been updated with a backwards-incompatible change (and thus the service cannot refer to the "normal" set up instructions).